requirements.txt
openai==0.27.6
langchain==0.0.167
python-dotenv==1.0.0
streamlit==1.22.0
streamlit-chat==0.0.2.2